+++ RabbitMQ.AMQP.Client/IPublisher.cs
@@ -2,6 +2,7 @@
 // and the Mozilla Public License, version 2.0.
 // Copyright (c) 2017-2024 Broadcom. All Rights Reserved. The term "Broadcom" refers to Broadcom Inc. and/or its subsidiaries.
 
+using System;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -38,10 +39,12 @@ namespace RabbitMQ.AMQP.Client
     /// </summary>
     public class PublishOutcome
     {
-        public PublishOutcome(OutcomeState state, Error? error)
+        public PublishOutcome(OutcomeState state, Error? error,
+            AmqpMessageRejectedException? exception = null)
         {
             State = state;
             Error = error;
+            Exception = exception;
         }
 
         /// <summary>
@@ -53,6 +56,12 @@ namespace RabbitMQ.AMQP.Client
         /// The <see cref="Error"/>, if any.
         /// </summary>
         public Error? Error { get; }
+
+        /// <summary>
+        /// An <see cref="AmqpMessageRejectedException"/> if the message was rejected and the broker
+        /// provided rejection details (requires RabbitMQ 4.3 or later). <c>null</c> otherwise.
+        /// </summary>
+        public AmqpMessageRejectedException? Exception { get; }
     }
 
     /// <summary>
@@ -78,6 +87,16 @@ namespace RabbitMQ.AMQP.Client
     /// </summary>
     public interface IPublisher : ILifeCycle
     {
+        /// <summary>
+        /// Publishes a message to the broker asynchronously.
+        /// </summary>
+        /// <param name="message"></param>
+        /// <param name="cancellationToken"></param>
+        /// <returns>A <see cref="Task"/> representating the await-able result of the publish operation.</returns>
+        /// <exception cref="InvalidOperationException"></exception>
+        /// <exception cref="NotSupportedException"></exception>
+        /// <exception cref="PublisherException"></exception>
+
         Task<PublishResult> PublishAsync(IMessage message, CancellationToken cancellationToken = default);
     }
 }